# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SoundBoard - Community Music PWA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List songs in library
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: storageType
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Song list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"
    post:
      operationId: createSong
      tags: [songs]
      summary: Add a song to the library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
  /songs/{id}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get a song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          description: Not found
    patch:
      operationId: updateSong
      tags: [songs]
      summary: Update song metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /songs/{id}/promote:
    post:
      operationId: promoteSong
      tags: [songs]
      summary: Promote song from limited to permanent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteInput"
      responses:
        "200":
          description: Promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"

  # ── Search ─────────────────────────────────────────────────────────────────
...
  /songs/{id}/tags:
    post:
      operationId: tagSong
      tags: [library]
      summary: Add a mood/category tag to a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "200":
          description: Tagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"

  # ── Playlists ──────────────────────────────────────────────────────────────
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register or update a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/inbox:
    get:
      operationId: getUserInbox
      tags: [users]
      summary: Get user inbox (shared songs as gifts)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Inbox items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboxItem"
  /users/share:
    post:
      operationId: shareSong
      tags: [users]
      summary: Share a song with another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareInput"
      responses:
        "201":
          description: Shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"

  # ── Stats / Dashboard ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Song:
      type: object
      required: [id, title, artist, source, sourceUrl, storageType, category, userId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        album:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        coverUrl:
          type: ["string", "null"]
        source:
          type: string
          enum: [youtube, soundcloud, upload]
        sourceUrl:
          type: string
        storageType:
          type: string
          enum: [limited, permanent, public_limited, public_download]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        userId:
          type: string
        expiresAt:
          type: ["string", "null"]
        voteCount:
          type: integer
        isPublic:
          type: boolean
        createdAt:
          type: string
    SongInput:
      type: object
      required: [title, artist, source, sourceUrl, storageType, userId]
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration:
          type: integer
        coverUrl:
          type: string
        source:
          type: string
          enum: [youtube, soundcloud, upload]
        sourceUrl:
          type: string
        storageType:
          type: string
          enum: [limited, permanent, public_limited, public_download]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        userId:
          type: string
        isPublic:
          type: boolean
    SongUpdate:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        coverUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        storageType:
          type: string
          enum: [limited, permanent, public_limited, public_download]
    PromoteInput:
      type: object
      required: [targetStorage]
      properties:
        targetStorage:
          type: string
          enum: [permanent, public_download]
    TagInput:
      type: object
      required: [tag, userId]
      properties:
        tag:
          type: string
        userId:
          type: string

    # ── Search ─────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [userId, displayName, reputation, badgeCount, createdAt]
      properties:
        userId:
          type: string
        displayName:
          type: string
        reputation:
          type: integer
        badgeCount:
          type: integer
        badges:
          type: array
          items:
            type: string
        nominatedCount:
          type: integer
        uploadCount:
          type: integer
        createdAt:
          type: string
    UserInput:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
    InboxItem:
      type: object
      required: [id, fromUserId, toUserId, songId, song, createdAt]
      properties:
        id:
          type: integer
        fromUserId:
          type: string
        fromUserName:
          type: ["string", "null"]
        toUserId:
          type: string
        songId:
          type: integer
        song:
          $ref: "#/components/schemas/Song"
        message:
          type: ["string", "null"]
        createdAt:
          type: string
    ShareInput:
      type: object
      required: [fromUserId, toUserId, songId]
      properties:
        fromUserId:
          type: string
        toUserId:
          type: string
        songId:
          type: integer
        message:
          type: string

    # ── Stats ──────────────────────────────────────────────────────────────
...