# lib/api-zod/src/generated/types/inboxItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundBoard - Community Music PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface InboxItem {
  id: number;
  fromUserId: string;
  /** @nullable */
  fromUserName?: string | null;
  toUserId: string;
  songId: number;
  song: Song;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}