# lib/api-zod/src/generated/types/promoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundBoard - Community Music PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { PromoteInputTargetStorage } from './promoteInputTargetStorage';

export interface PromoteInput {
  targetStorage: PromoteInputTargetStorage;
}