# lib/api-zod/src/generated/types/shareInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundBoard - Community Music PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareInput {
  fromUserId: string;
  toUserId: string;
  songId: number;
  message?: string;
}