# lib/api-zod/src/generated/types/songInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundBoard - Community Music PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { SongInputSource } from './songInputSource';
import type { SongInputStorageType } from './songInputStorageType';

export interface SongInput {
  title: string;
  artist: string;
  album?: string;
  duration?: number;
  coverUrl?: string;
  source: SongInputSource;
  sourceUrl: string;
  storageType: SongInputStorageType;
  category?: string;
  tags?: string[];
  userId: string;
  isPublic?: boolean;
}