# lib/api-zod/src/generated/types/songUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundBoard - Community Music PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { SongUpdateStorageType } from './songUpdateStorageType';

export interface SongUpdate {
  title?: string;
  artist?: string;
  album?: string;
  coverUrl?: string;
  category?: string;
  tags?: string[];
  storageType?: SongUpdateStorageType;
}